+++ application/app/Http/Controllers/TaskController.php
@@ -12,7 +12,6 @@ use App\Models\TaskPicture;
 use App\Models\User;
 use Illuminate\Http\Request;
 use App\Http\Requests\TaskStoreRequest;
-use App\Http\Requests\TaskPictureStoreRequest;
 use App\Http\Requests\TaskUpdateRequest;
 
 class TaskController extends Controller
@@ -80,7 +79,7 @@ class TaskController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create(Request $request, Project $project)
+    public function create(Project $project)
     {
         $task_kinds = TaskKind::all();
         $task_statuses = TaskStatus::all();
