+++ application/app/Http/Controllers/TmpPictureController.php
@@ -5,29 +5,34 @@ namespace App\Http\Controllers;
 use App\Models\Project;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Str;
 use App\Http\Requests\TaskPictureStoreRequest;
 
 class TmpPictureController extends Controller
 {
     /**
-     * Store a newly created resource in storage.
+     * store the picture temporarily.
      *
-     * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request, Project $project)
+    public function storeTmpPicture(TaskPictureStoreRequest $request, Project $project)
     {
         //投稿予定画像がある場合、セッション情報を一時的に保持する
         if($file = $request->file('file')){
-            $client_original_name = $file->getClientOriginalName();
+            //キー名としてランダムな名前を付与
+            $tmp_file_name = Str::random(20);
+            //storage/app/public/tmpにファイルを保存し、パスを取得
             $tmp_file_path = basename($file->store('public/tmp'));
-            $request->session()->push('tmp_files.path', $tmp_file_path);
-            $request->session()->push('tmp_files.name', $client_original_name);
-            // ddd($file, $request->session());
+            $request->session()->put('tmp_files.'.$tmp_file_name, $tmp_file_path);
+
+            $flash = ['success' => __('Picture added successfully.')];
+        } else {
+            $flash = ['error' => __('Failed to add the picture.')];
         }
 
-        return redirect()->route('tasks.create', ['project' => $project->id]);
-
+        return redirect()
+            ->route('tasks.create', ['project' => $project->id])
+            ->with($flash);
     }
 
     /**
@@ -36,17 +41,20 @@ class TmpPictureController extends Controller
      * @param  \App\Models\TaskPicture  $taskPicture
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Project $project, Request $tmp_stored_picture_path)
+    public function destroyTmpPicture(Project $project, Request $request)
     {
-        if (session()->has('tmp_stored_picture_path', $tmp_stored_picture_path)) {
-            Storage::disk('public/tmp')->delete($tmp_stored_picture_path);
-            session()->forget('tmp_stored_picture_path',$tmp_stored_picture_path);
+        $tmp_file_name = $request->input('tmp_file_name');
+        if (session()->has('tmp_files.'.$tmp_file_name)) {
+            //セッションからパスを取り出してファイルを削除
+            $tmp_file_path = $request->session()->pull('tmp_files.'.$tmp_file_name);
+            Storage::disk('public')->delete('tmp/'.$tmp_file_path);
+
             $flash = ['success' => __('Picture deleted successfully.')];
         } else {
             $flash = ['error' => __('Failed to delete the picture.')];
         }
-
-        return redirect()->route('tasks.create', ['project' => $project->id])
+        return redirect()
+            ->route('tasks.create', ['project' => $project->id])
             ->with($flash);
     }
 }
